// frontend/src/api/generated/maia.zod.ts
export const moveResponseTopMovesItemWdlMin = 3;
export const moveResponseTopMovesItemWdlMax = 3;

export const moveResponseWdlMin = 3;
export const moveResponseWdlMax = 3;



export const MoveResponse = zod.strictObject({
  "$schema": zod.url().optional().describe('A URL to the JSON Schema for this object.'),
  "degraded": zod.boolean(),
  "delta_baseline": zod.strictObject({
  "kind": zod.string(),
  "value": zod.number()
}).optional(),
  "model_used": zod.string(),
  "move": zod.string(),
  "top_moves": zod.array(zod.strictObject({
  "delta": zod.number().optional(),
  "move": zod.string(),
  "prob": zod.number(),
  "wdl": zod.array(zod.number()).min(moveResponseTopMovesItemWdlMin).max(moveResponseTopMovesItemWdlMax).nullable()
})).nullable(),
  "wdl": zod.array(zod.number()).min(moveResponseWdlMin).max(moveResponseWdlMax).nullable()
})